move broadly useful fns to src/common/utils.js
