+++ src/peer/util.js
@@ -7,6 +7,7 @@ const PeerInfo = require('peer-info')
 const lp = require('pull-length-prefixed')
 const {decode} = require('../metadata/serialize')
 const _ = require('lodash')
+const { flatMap } = require('../common/util')
 
 import type { PeerInfoMsg, LookupPeerResponseMsg, ProtoCodec, QueryResultMsg, QueryResultValueMsg, SimpleValueMsg, DataResultMsg, DataObjectMsg, StatementMsg, StatementBodyMsg, SimpleStatementMsg } from '../protobuf/types'  // eslint-disable-line no-unused-vars
 
@@ -165,24 +166,6 @@ const resultStreamThrough: MediachainStreamThrough<*> = (read) => {
   }
 }
 
-/**
- * Reject `promise` if it doesn't complete within `timeout` milliseconds
- * @param timeout milliseconds to wait before rejecting
- * @param promise a promise that you want to set a timeout for
- * @returns a Promise that will resolve to the value of `promise`, unless the timeout is exceeded
- */
-function promiseTimeout<T> (timeout: number, promise: Promise<T>): Promise<T> {
-  return Promise.race([promise, new Promise((resolve, reject) => {
-    setTimeout(() => {
-      reject(new Error(`Timeout of ${timeout}ms exceeded`))
-    }, timeout)
-  })])
-}
-
-function flatMap<T, U> (array: Array<T>, f: (x: T) => Array<U>): Array<U> {
-  return [].concat(...array.map(x => f(x)))
-}
-
 function objectIdsForQueryResult (result: QueryResultValueMsg): Array<string> {
   let values: Array<SimpleValueMsg> = []
 
@@ -258,8 +241,6 @@ module.exports = {
   pullToPromise,
   pullRepeatedly,
   resultStreamThrough,
-  promiseTimeout,
   objectIdsForQueryResult,
-  expandQueryResult,
-  flatMap
+  expandQueryResult
 }
